requirements.txt
lxml==4.2.1
requests==2.21.0
jsonpickle==1.0
python_dateutil==2.7.3
